App/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
pyjwt
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
pydantic-settings
pandas
fpdf
openpyxl
reportlab
jinja2